fix log;

+++ test/utest_get_mcu_data.cpp
@@ -73,8 +73,8 @@ void print_McuKey_st(const McuKey_st &mcu_key) {
   std::ostringstream oss;
 
   oss << "info McuKey_st: " << __func__ << "\n"
-      << "Key = " << mcu_key.Key << "\n"
-      << "battery_percent = " << mcu_key.battery_percent << "\n";
+      << "Key = " << static_cast<uint32_t>(mcu_key.Key) << "\n"
+      << "battery_percent = " << static_cast<uint32_t>(mcu_key.battery_percent) << "\n";
 
   std::cout << oss.str() << std::endl;
 }
